Emit each function in a separate section (amd64,i386,arm,arm64)
With this option, the compiler will emit each function in a separate
named text section, on supported targets. This enables function
reordering using a linker script. With this option, the compiler also
emits caml_hot__code_begin and caml_hot__code_end sections. This
allows a linker script to move function sections outside of the
segments they belong to, without breaking caml_code_segments.

This configure option is enabled by default. The compiler itself is
built with this option enabled, by default.  To emit function sections
for user code, compile-time option -function-sections is required.

If the user explicitly requests function sections by calling configure
with --enable-function-sections on a target that does not support it,
it's an error at configure time.