Revised handling of weights over preference edges (#8909)
Before register allocation we build a graph of interferences and preferences between pseudo-registers.  Preferences carry an integer weight to reflect the idea that some preferences are "more important" than others and should be honored first.

Currently, we divide the weight by 2 when processing the branches of a conditional.
- This is problematic because weight can drop to 0 and then preferences are ignored.
- This does not improve performance, and might actually reduce the effectiveness of coalescing.

This PR solves both problems at once by not reducing the weight in branches.  Consequently, the weight never drops to 0.