ocamltest: fix recursive expansion of variables
ocamltest uses Buffer.add_substitute to perform variable substitution.
The function also replaces escaped `$` characters (i.e., the string
`\$`) with the character `$`.
ocamltest would apply substitution in a loop until the input and
substituted strings were equal; however, removing the escape character
during the first iteration would create variables during the second
iteration.
Consider now that a fixpoint is reached when no substitutions have
been performed.

Other problems occur when substituting on the whole string: the number
of substitutions on the whole string is determined by the maximum
number of recursive expansions of a single variable, should it expand
in turn to another variable. With this algorithm, a variable that
should be expanded only once would be expanded at most as many times
as the variable with the maximum number of expansions of the original
string, thus exhibiting inconsistent behavior.

The solution is to apply substitution recursively only on the expanded
values, not always on the whole string, and to keep track as above if
substitutions have been performed.